--- README.md
@@ -1,8 +1,0 @@
-# portfolio-project
-Pour la semaine de projet React/Node js nous avons opté pour un portfolio dynamique.
-
-Je suis un test numero 2
-Tai : je suis connecté merci =)
-
-Tai est passé par là
-test création branch Tai
