+++ frontend/src/pages/Home.module.css
@@ -1,0 +1,6 @@
+.divContainer{
+    display: flex;
+    justify-content: space-around;
+    flex-wrap: wrap;
+
+}
