--- myprojects/src/App.js
@@ -1,12 +1,0 @@
-import './App.css';
-import Home from './pages/Home';
-
-function App() {
-  return (
-    <div>
-      <Home/>
-    </div>
-  )
-}
-
-export default App;
