--- myprojects/src/components/Cards/Cards.module.css
@@ -1,27 +1,0 @@
-.container{
-    display: flex;
-    flex-direction: row;
-    justify-content: space-around;
-
-}
-
-.cards{
-    background-color: grey;
-    width: 300px;
-    height: 350px;
-    margin: 20px;
-}
-.cercle{
-    background-color: white;
-    width: 200px;
-    height: 200px;
-    border-radius: 50%;
-    margin-left: 50px;
-    margin-top: 20px;
-}
-.commentBox{
-    background-color: white;
-    width: 280px;
-    height: 90px;
-    margin: 20px 10px;
-}
