--- myprojects/src/components/Nav/Nav.module.css
@@ -1,16 +1,0 @@
-.navbar{
-    color: white;
-    background-color: #323232;
-}
-
-.imgLogo{
-    width: 40px;
-    height: 40px;
-}
-
-.menus{
-    display: flex;
-    justify-content: flex-end;
-    list-style-type: none;
-    margin: 5px;
-}
