--- myprojects/src/pages/Home.jsx
@@ -1,19 +1,0 @@
-import React from 'react';
-import Cards from '../components/Cards/Cards';
-import Nav from '../components/Nav/Nav';
-import Carousel from '../components/Carousel/Carousel';
-
-function Home() {
-    return (
-        <div>
-            <Nav/>
-            <Carousel/>
-            <Cards />
-
-                        
-
-        </div>
-    )
-}
-
-export default Home
