--- src/components/Cards/Card.module.css
@@ -1,9 +1,0 @@
-.media{
-    height: 140,
-  }
-  .dimension{
-      margin: 80px;
-      height: 250px;
-      width: 250px;
-      display: flex;
-  }
