--- src/pages/Home.module.css
@@ -1,6 +1,0 @@
-.divContainer{
-    display: flex;
-    justify-content: space-around;
-    flex-wrap: wrap;
-
-}
